+++ src/ais/index.ts
@@ -2,6 +2,7 @@ import { AIBase } from "./_base";
 import { AmazonsAI } from "./amazons";
 import { BlamAI } from "./blam";
 import { CannonAI } from "./cannon";
+import { MchessAI } from "./mchess";
 
 export interface IAIResult {
     bestMove: string|null;
@@ -12,17 +13,19 @@ export interface IAI {
     findmove: (state: any, depth: number) => string;
 }
 
-export { AIBase, AmazonsAI, BlamAI, CannonAI };
+export { AIBase, AmazonsAI, BlamAI, CannonAI, MchessAI };
 
 export const supportedGames: string[] = ["amazons"];
 export const fastGames: Map<string, number> = new Map([
     ["amazons", 1],
     ["blam", 3],
-    ["cannon", 3]
+    ["cannon", 3],
+    ["mchess", 5]
 ]);
 export const slowGames: Map<string, number> = new Map([
     ["blam", 5],
-    ["cannon", 5]
+    ["cannon", 5],
+    ["mchess", 7]
 ]);
 
 export function AIFactory(game: string): AIBase|undefined {
@@ -33,6 +36,8 @@ export function AIFactory(game: string): AIBase|undefined {
             return new BlamAI();
         case "cannon":
             return new CannonAI();
+        case "mchess":
+            return new MchessAI();
     }
     return;
 }